refactor(core): repackage semantic build
Why:
- Core semantic extraction was scattered across flat `application/core/*` files, which blurred the boundary between semantic-build ownership and downstream Core consumers.
- Issue #90 tracks the packaging cleanup needed to make the Source-to-Core semantic build path explicit under the structured extraction cutover.

What:
- Move document profile, paper facts, and semantic version modules into `application/core/semantic_build/`.
- Move Core-owned prompt, schema, and extractor files into `application/core/semantic_build/llm/` and repoint direct imports across Core consumers, controllers, source task runner, and tests.
- Add semantic-build package docs and update plan-doc links so docs governance matches the new package layout.

Impact:
- Core semantic build now reads as one explicit package without compatibility shims or dual-path imports.
- Runtime behavior is intended to stay the same; the change is an internal ownership and maintainability cleanup.

Tests:
- `python3 -m py_compile backend/application/core/comparison_service.py backend/application/core/workspace_overview_service.py backend/application/source/collection_build_task_runner.py backend/application/core/semantic_build/document_profile_service.py backend/application/core/semantic_build/paper_facts_service.py backend/application/core/semantic_build/core_semantic_version.py backend/application/core/semantic_build/llm/extractor.py backend/application/core/semantic_build/llm/prompts.py backend/application/core/semantic_build/llm/schemas.py backend/controllers/core/documents.py backend/controllers/core/evidence.py backend/tests/conftest.py backend/tests/support/fake_core_llm_extractor.py backend/tests/unit/services/test_document_profile_service.py backend/tests/unit/services/test_paper_facts_services.py backend/tests/unit/services/test_core_semantic_version.py`
- `python3 scripts/check_docs_governance.py`
- `cd backend && .venv/bin/python -m pytest tests/unit/services/test_document_profile_service.py tests/unit/services/test_paper_facts_services.py tests/unit/services/test_core_semantic_version.py tests/unit/services/test_workspace_service.py`
- `cd backend && .venv/bin/python -m pytest tests/unit/routers/test_documents_api.py tests/unit/routers/test_evidence_api.py tests/unit/routers/test_comparisons_api.py tests/integration/test_app_layer_api.py -q -rs` (skipped: `fastapi` not installed)

Refs:
- #90
- Parent: #84
