+++ backend/application/source/collection_build_task_runner.py
@@ -9,8 +9,8 @@ from infra.source.config.pipeline_mode import IndexingMethod
 
 from application.core.comparison_service import ComparisonService
 from application.source.collection_service import CollectionService
-from application.core.document_profile_service import DocumentProfileService
-from application.core.paper_facts_service import PaperFactsService
+from application.core.semantic_build.document_profile_service import DocumentProfileService
+from application.core.semantic_build.paper_facts_service import PaperFactsService
 from application.source.task_service import TaskService
 from application.derived.protocol.pipeline_service import build_protocol_artifacts
 from application.source.artifact_registry_service import ArtifactRegistryService
