+++ backend/controllers/core/documents.py
@@ -4,7 +4,7 @@ from typing import Annotated
 
 from fastapi import APIRouter, HTTPException, Query
 
-from application.core.document_profile_service import (
+from application.core.semantic_build.document_profile_service import (
     DocumentContentNotReadyError,
     DocumentNotFoundError,
     DocumentProfileService,
