+++ backend/tests/conftest.py
@@ -12,7 +12,11 @@ if str(ROOT) not in sys.path:
 
 @pytest.fixture(autouse=True)
 def _patch_core_llm_extractor(monkeypatch):
-    from application.core import comparison_service, document_profile_service, paper_facts_service
+    from application.core import comparison_service
+    from application.core.semantic_build import (
+        document_profile_service,
+        paper_facts_service,
+    )
     from tests.support.fake_core_llm_extractor import FakeCoreLLMStructuredExtractor
 
     fake = FakeCoreLLMStructuredExtractor()
